Cut down AccelEngine.
